[HW-2-4-1] переделал linkedlist
